+++ 1-ContenuProjet/GestTask/GestTask/GestTask/ViewModels/TaskMenuViewModel.cs
@@ -1,4 +1,9 @@
-﻿using GestTask.Models;
+﻿/* Developper : Tristan Gerber
+ * Place : ETML, N501
+ * Project creation date : 05.05.2022
+ * Last updated : 25.05.2022 */
+
+using GestTask.Models;
 using Rg.Plugins.Popup.Contracts;
 using Rg.Plugins.Popup.Services;
 using System;
@@ -14,7 +19,7 @@ namespace GestTask.ViewModels
         private string name;
         private string description;
         private bool inToDoList;
-        private bool active;
+        private bool finished;
         private ObservableCollection<CategoryModel> categories;
         private CategoryModel selectedCategory;
         private TasksViewModel _baseViewModel;
@@ -22,7 +27,7 @@ namespace GestTask.ViewModels
         public string Name { get => name; set => SetProperty(ref name, value); }
         public string Description { get => description; set => SetProperty(ref description, value); }
         public bool InToDoList { get => inToDoList; set => SetProperty(ref inToDoList, value); }
-        public bool Active { get => active; set => SetProperty(ref active, value); }
+        public bool Finished { get => finished; set => SetProperty(ref finished, value); }
         public ObservableCollection<CategoryModel> Categories { get => categories; set => SetProperty(ref categories, value); }
         public CategoryModel SelectedCategory { get => selectedCategory; set => SetProperty(ref selectedCategory, value); }
 
@@ -40,7 +45,7 @@ namespace GestTask.ViewModels
             Name = task.Name;
             Description = task.Description;
             InToDoList = task.InToDoList;
-            Active = task.Active;
+            Finished = task.Finished;
             Categories = App.Db.GetCategoriesAsync();
             SelectedCategory = App.Db.GetCategoryAsync(task.FkCategory).Result;
 
@@ -52,12 +57,14 @@ namespace GestTask.ViewModels
 
         private async Task ExecuteDeleteCommand()
         {
-            
-            await App.Db.DeleteTaskAsync(_task);
+            if (await App.Current.MainPage.DisplayAlert("Confirmation", "Êtes vous sur de vouloir supprimer ?", "Oui", "Non"))
+            {
+                await App.Db.DeleteTaskAsync(_task);
 
-            // Navigate backwards
-            _baseViewModel.ExecuteLoadTasksCommand();
-            await _popup.PopAsync();
+                // Navigate backwards
+                _baseViewModel.ExecuteLoadTasksCommand();
+                await _popup.PopAsync();
+            }
         }
 
         private async Task ExecuteSaveCommand()
@@ -66,12 +73,23 @@ namespace GestTask.ViewModels
             _task.Name = name;
             _task.Description = description;
             _task.InToDoList = inToDoList;
-            _task.Active = active;
-            _task.FkCategory = SelectedCategory.Id;
+            _task.Finished = finished;
+            if (_task.Finished)
+            {
+                _task.InToDoList = false;
+            }
+            if (selectedCategory != null)
+            {
+                _task.FkCategory = selectedCategory.Id;
+            }
 
             if (!string.IsNullOrWhiteSpace(_task.Name))
             {
                 await App.Db.SaveTaskAsync(_task);
+            }
+            else
+            {
+                await App.Current.MainPage.DisplayAlert("Erreur", "Veuillez remplir le nom de la tâche", "Retour");
             }
 
             // Navigate backwards
